fc-project-board/src/main/java/com/fc/projectboard/service/ArticleService.java
package com.fc.projectboard.service;

import com.fc.projectboard.domain.type.SearchType;
import com.fc.projectboard.dto.ArticleDto;
import com.fc.projectboard.dto.ArticleUpdateDto;
import com.fc.projectboard.repository.ArticleRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleService {

    private final ArticleRepository articleRepository;

    @Transactional(readOnly = true)
    public Page<ArticleDto> searchArticles(SearchType title, String searchKeyword) {
        return Page.empty();
    }

    @Transactional(readOnly = true)
    public ArticleDto searchArticle(long l) {
        return null;
    }

    public void saveArticle(ArticleDto dto) {
        // TODO document why this method is empty
    }

    public void updateArticle(long articleId, ArticleUpdateDto dto) {
        // TODO document why this method is empty
    }

    public void deleteArticle(long articleId) {
        // TODO document why this method is empty
    }
}
